Split apps/users/views into endpoint-specific modules
Replace apps/users/views/_impl.py (683 lines) and apps/users/views/profile.py
(21 lines) with five focused modules: account (current user / stats), avatar,
login_records, password (change / forgot / reset), and preferences. Public
view names are unchanged; views/__init__.py keeps the same exports.
