--- backend/apps/users/views/profile.py
@@ -1,21 +1,0 @@
-"""Current-user profile and preferences views."""
-
-from ._impl import (
-    CurrentUserView,
-    UserStatsView,
-    UserPreferencesView,
-    UserAvatarUploadView,
-    ChangePasswordView,
-    ForgotPasswordView,
-    ResetPasswordView,
-)
-
-__all__ = [
-    "CurrentUserView",
-    "UserStatsView",
-    "UserPreferencesView",
-    "UserAvatarUploadView",
-    "ChangePasswordView",
-    "ForgotPasswordView",
-    "ResetPasswordView",
-]
